# src/main/java/com/jh/tds/as/model/AuthRequest.java
// AuthRequest.java
package com.jh.tds.as.model;

public class AuthRequest {
    private String username;
    private String password;
    private String role;

    // Getters and Setters

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public String getRole() {
        return role;
    }

    public void setRole(String role) {
        this.role = role;
    }
}