# src/main/java/com/jh/tds/as/model/AuthResponse.java
package com.jh.tds.as.model;

/**
 * @author Veeresh N
 * @version 1.0
 */
public class AuthResponse {
    private String token;

    public String getToken() {
        return token;
    }

    public void setToken(String token) {
        this.token = token;
    }
}